[DAGE-43] added unification of query_template syntax
Now we give the path and validates it. If we don't provide the query_template_path param in the config.yaml file it skips the second retrieval (after the generation of the queries). rre_query_template_path is kept so we support 2 different search. If not provided, is set to the query_template_path one. If both of them are not provided and we are looking to save in RRE format, it raises and error.