[DAGE-74] fixed bug in solr-init.sh
The old version was indexing just the documents that were found in the documents_embeddings.jsonl file.
The new version it indexes all the documents, even if some embedding is not found. In this case, the collection will have some documents without embeddings, but all the available ones are indexed.