[DAGE-88] embedding_writer.py is now fixed
The problem was the following: the preprocessing was good for reranking_task.py but NOT for retrieval_task.py. Looking at class `AbsTaskRetrieval` it can be spotted that the structure of the string to send to the evaluation process is different. Now we have 2 different function in helper.py, that have been separated for the 2 different tasks. This solves the problem of duplicated embeddings in the cache (can be spotted by using the logger in DEBUG mode -> the documents are pushed 2 times).
Another problem came out from logs in DEBUG mode: when I set the reranking task, the cache inside has only the documents that has a score in the candidates.jsonl file. This behaviour makes a lot of sense, since we don't need to have the embeddings of the whole corpus to compute the metrics needed. Since, in the documents_embeddings.jsonl file, we want the embedding of all the docs inside corpus.jsonl file, we still need to remove name and normalize_embeddings, as in Naz PR. The embeddings of the documents which does NOT appear in candidates, are computed by the SentenceTransformer model linked to the cache while the file is being written in the `resource/embeddings/` folder.